refactor: extract SQL enum serialization to macro to reduce duplication
Eliminate byte-for-byte identical FromSql/ToSql implementations across
5 enum types (AgentRunStatus, FeedbackStatus, StepStatus, WorkflowRunStatus,
WorkflowStepStatus) by introducing impl_sql_enum! macro.

All 5 types already implement Display + FromStr<Err = String>, so the
macro generates both ToSql and FromSql implementations that parse from
string via the FromStr trait, reducing ~85 lines of boilerplate to 5
macro invocations.
